requirements.txt
python-socketio
requests